move dem files
