+++ client/containers/DashboardSubmissions/shared.scss
