+++ client/containers/DashboardSubmissions/submissionItems.scss
@@ -1,4 +1,4 @@
-@import '../overviewStyles.scss';
+@import '../DashboardOverview/overviewStyles.scss';
 
 .submission-items-component {
     .bp3-non-ideal-state {
